Use MUnit in TypedQuerySuite
